Update 4.2_Variant_Selection.R
I guess what you meant is to get a vector consisting of TRUE or FALSE here, right? The current expression would get tons of duplicated cells. It won't affect any of the downstream analysis, but since I was just repeating your work, I thought it would be better that we correct this tiny error.